move function to utils file
